requirements.txt
torch
torchvision
transformers
pdf2image
pillow
pandas
streamlit
rich
